requirements.txt
python-telegram-bot==12
requests==2.21.0
emoji==0.5.1
